pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "apicorex"
version = "0.1.0"
description = "A Python API Automation Framework"
authors = [
    { name = "Aishwarya", email = "author@example.com" }
]
requires-python = ">=3.8"
dependencies = [
    "requests==2.31.0",
    "pytest==8.1.1",
    "pytest-html==4.1.1",
    "allure-pytest==2.13.5",
    "python-dotenv==1.0.1",
    "selenium>=4.10.0",
    "Appium-Python-Client>=3.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]
